# 8266528: Optimize C2 VerifyIterativeGVN execution time