Optimization for VerifyIterativeGVN, motiviation is running with -XX:+VerifyIterativeGVN is extremly slow.

In simple test "-Xcomp -XX:+VerifyIterativeGVN -XX:-TieredCompilation -version", time reduced from 8.67s to 2.4s.
In extreme case hotspot/test/jtreg/compiler/escapeAnalysis/Test6689060.java, time reduced from 20000s to 92s.
Detail data in JBS description.

Optimizations includes:
1. Optimize redundant verfications in PhaseIterGVN::verify_step. Nodes might verified multiple times.
    Redundant verifications between full pass and _verify_window single node process.
    Redundant verifications between different nodes in _verify_window

2. Optimize def-use edge checking: 
    Skip multiple checks for same x->n input edges.
    Skip redundant check in inner loop when counting how many x in n's input edges, skip current index.

3. Optimize field access
    Replace "n->in(j)" with "n->_in[j]", skipping unuseful assert when invoking Node::in(int index).

Optimization#2/#3 decrease execution time and no other overhead. 
optimization#1 adds 3 fields in class Node in debug build,  they can be squeezed into an "int/long" if needed.
```
  jint _igvn_verify_depth_cur;
  jint _igvn_verify_depth_prev;
  julong _igvn_verify_epoch;
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [ ] Change must be properly reviewed

### Issue
 * [JDK-8266528](https://bugs.openjdk.java.net/browse/JDK-8266528): Optimize C2 VerifyIterativeGVN execution time


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/3872/head:pull/3872` \
`$ git checkout pull/3872`

Update a local copy of the PR: \
`$ git checkout pull/3872` \
`$ git pull https://git.openjdk.java.net/jdk pull/3872/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 3872`

View PR using the GUI difftool: \
`$ git pr show -t 3872`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/3872.diff">https://git.openjdk.java.net/jdk/pull/3872.diff</a>

</details>