@huishi-hs 2021-05-05
Not sure if adding fields in Node class debug builds in acceptable for verification purpose. Would appreciate for any comments.